+++ lib/features/auth/presentation/screens/signup_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:fruits_hub/core/services/get_it_service.dart';
 import 'package:fruits_hub/core/widgets/custom_app_bar.dart';
-import 'package:fruits_hub/features/auth/presentation/cubits/signup_cubit.dart';
+import 'package:fruits_hub/features/auth/presentation/cubits/signup/signup_cubit.dart';
 import 'package:fruits_hub/features/auth/presentation/widgets/signup/signup_body_bloc_consumer.dart';
 import 'package:fruits_hub/generated/l10n.dart';
