+++ lib/features/auth/presentation/widgets/signup/signup_body_bloc_consumer.dart
@@ -4,8 +4,8 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:fruits_hub/core/enums/enums.dart';
 import 'package:fruits_hub/core/helper/extensions.dart';
 import 'package:fruits_hub/core/helper/show_toast.dart';
-import 'package:fruits_hub/features/auth/presentation/cubits/signup_cubit.dart';
-import 'package:fruits_hub/features/auth/presentation/cubits/signup_state.dart';
+import 'package:fruits_hub/features/auth/presentation/cubits/signup/signup_cubit.dart';
+import 'package:fruits_hub/features/auth/presentation/cubits/signup/signup_state.dart';
 import 'package:fruits_hub/features/auth/presentation/widgets/signup/signup_screen_body.dart';
 import 'package:fruits_hub/generated/l10n.dart';
 import 'package:modal_progress_hud_nsn/modal_progress_hud_nsn.dart';
